requirements.txt
seleniumbase
python-dotenv
flask
flask-login
flask-sqlalchemy
flask_basicauth
pymysql
python-crontab
beautifulsoup4
html5lib
lxml
mysql-connector
Flask-APScheduler